+++ src/services/hooks/useUpdate/useUpdate.ts
@@ -4,7 +4,17 @@ import { UpdateCallback } from '../types';
 
 export const useUpdate = (callback: UpdateCallback, dependencyList: DependencyList): void => {
   useEffect(() => {
-    callback();
+    const result = callback();
+
+    return () => {
+      (async () => {
+        const func = await result;
+
+        if (typeof func === 'function') {
+          func();
+        }
+      })();
+    };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, dependencyList);
 };