+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [0.1.3]() - 2022-05-01
+
+### Added
+
+- Add `useLastDiffValue`, `useResizeObserver` and `useWindowResize` tests.
+- Update `README` contents.
+
 # [0.1.2]() - 2022-04-26
 
 ### Added